api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "y-agent-api"
version = "0.4.0"
description = "REST API for y-agent"
requires-python = ">=3.10"
dependencies = [
    "y-agent-storage",
    "y-agent-agent",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sse-starlette>=2.0.0",
    "PyJWT>=2.8.0",
    "google-auth>=2.29.0",
    "requests>=2.31.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.uv.sources]
y-agent-storage = { workspace = true }
y-agent-agent = { workspace = true }
